+++ app/src/main/java/in/teacher/dao/ActivityMarkDao.java
@@ -26,9 +26,9 @@ public class ActivityMarkDao {
         return avg;
     }
 
-    public static List<String> selectActivityMarc(long activityId, List<Integer> studentId, SQLiteDatabase sqliteDatabase) {
+    public static List<String> selectActivityMarc(long activityId, List<Long> studentId, SQLiteDatabase sqliteDatabase) {
         List<String> mList = new ArrayList<>();
-        for (Integer i : studentId) {
+        for (Long i : studentId) {
             Cursor c = sqliteDatabase.rawQuery("select Mark from activitymark where ActivityId=" + activityId + " and StudentId=" + i, null);
             c.moveToFirst();
             if (c.getCount() > 0) {
@@ -159,7 +159,7 @@ public class ActivityMarkDao {
         }
     }
 
-    public static int getStudActAvg(int studentId, long activityId, SQLiteDatabase sqliteDatabase) {
+    public static int getStudActAvg(long studentId, long activityId, SQLiteDatabase sqliteDatabase) {
         int i = 0;
         Cursor c = sqliteDatabase.rawQuery("select (Avg(A.Mark)/B.MaximumMark)*100 as avg from activitymark A, activity B where A.ActivityId=B.ActivityId and A.ActivityId=" + activityId +
                 " and StudentId=" + studentId, null);
@@ -172,7 +172,7 @@ public class ActivityMarkDao {
         return i;
     }
 
-    public static int getStudActMark(int studentId, long activityId, SQLiteDatabase sqliteDatabase) {
+    public static int getStudActMark(long studentId, long activityId, SQLiteDatabase sqliteDatabase) {
         int i = 0;
         Cursor c = sqliteDatabase.rawQuery("select Mark from activitymark where StudentId=" + studentId + " and ActivityId=" + activityId, null);
         c.moveToFirst();
