+++ app/src/main/java/in/teacher/examfragment/UpdateSubActivityMark.java
@@ -69,7 +69,7 @@ public class UpdateSubActivityMark extends Fragment {
     private List<Students> studentsArray = new ArrayList<>();
     private List<Boolean> studentIndicate = new ArrayList<>();
     private ArrayList<Students> studentsArrayList = new ArrayList<>();
-    private List<Integer> studentsArrayId = new ArrayList<>();
+    private List<Long> studentsArrayId = new ArrayList<>();
     private List<String> studentScore = new ArrayList<>();
     private ListView lv;
     private MarksAdapter marksAdapter;
